Make sure VS2012 users receive proper update notifications
This commit changes the ID of the Editor Extensions VSIX to match the ID
used by the previous Code Contracts Editor Extensions VS2012 extension.
By using the same ID, existing users of this extension will receive and
upgrade notification after the next release of the editor extensions are
made available. Visual Studio 2010 users will not receive this
notification even though the extension can be installed in that
environment, but that is acceptable since Visual Studio 2010 did not
originally have a notification feature.